examples/etl_pipeline/flink_app/requirements.txt
pyflink==1.17.0
psycopg2-binary==2.9.6
